+++ changelog.md
@@ -1,5 +1,9 @@
 # changelog
 
+## 2.6.1
+* `FIX` modify luarc failed
+* `FIX` library files not recognized correctly
+
 ## 2.6.0
 `2022-1-13`
 * `NEW` supports multi-workspace in server side, for developers of language clients, please [read here](https://github.com/sumneko/lua-language-server/wiki/Multi-workspace-supports) to learn more.